+++ hatyan/analysis_prediction.py
@@ -8,7 +8,7 @@ import pandas as pd
 import datetime as dt
 import logging
 
-from hatyan.hatyan_core import get_const_list_hatyan, sort_const_list, robust_timedelta_sec, get_tstart_tstop_tstep
+from hatyan.hatyan_core import get_const_list_hatyan, sort_const_list, robust_timedelta_sec
 from hatyan.hatyan_core import get_freqv0_generic, get_uf_generic
 from hatyan.timeseries import Timeseries_Statistics, nyquist_folding, check_rayleigh
 from hatyan.metadata import metadata_from_obj, metadata_add_to_obj
@@ -273,7 +273,7 @@ def analysis_singleperiod(ts, const_list, hatyan_settings):
     bool_ts_duplicated = ts.index.duplicated(keep='first')
     ts_pd = ts.copy() #TODO: this is not necessary
     if bool_ts_duplicated.any():
-        raise Exception(f'ERROR: {bool_ts_duplicated.sum()} duplicate timesteps in provided timeseries, remove them e.g. with: ts = ts[~ts.index.duplicated(keep="first")]')
+        raise ValueError(f'{bool_ts_duplicated.sum()} duplicate timesteps in provided timeseries, remove them e.g. with: ts = ts[~ts.index.duplicated(keep="first")]')
     message = (f'#timesteps    = {len(ts)}\n'
                f'tstart        = {ts.index[0].strftime("%Y-%m-%d %H:%M:%S")}'
                f'tstop         = {ts.index[-1].strftime("%Y-%m-%d %H:%M:%S")}')
@@ -403,7 +403,7 @@ def split_components(comp, dood_date_mid, hatyan_settings):
     return comp_inclCS
 
 
-def prediction_singleperiod(comp:pd.DataFrame, times:(pd.DatetimeIndex,slice), hatyan_settings) -> pd.DataFrame:
+def prediction_singleperiod(comp:pd.DataFrame, times:pd.DatetimeIndex, hatyan_settings) -> pd.DataFrame:
     
     metadata_comp = metadata_from_obj(comp)
     tzone_comp = metadata_comp.pop("tzone")
@@ -471,7 +471,7 @@ def prediction_singleperiod(comp:pd.DataFrame, times:(pd.DatetimeIndex,slice), h
 
 
 @deprecated_python_option(**DEPRECATED_OPTIONS_PREDICTION_DICT)
-def prediction(comp, timestep_min=None, times=None):
+def prediction(comp, times=None, timestep=None):
     """
     generates a tidal prediction from a set of components A and phi values.
     The component set has the same timezone as the timeseries used to create it, 
@@ -483,12 +483,13 @@ def prediction(comp, timestep_min=None, times=None):
     ----------
     comp : pd.DataFrame
         The DataFrame contains the component data with component names as index, and colums 'A' and 'phi_deg'.
-    timestep_min : int
-        Only allowed/relevant for component dataframes with multi-level columns (different periods). The default is None.
     times : (pd.DatetimeIndex,slice), optional
         pd.DatetimeIndex with prediction timeseries or slice(tstart,stop,timestep) to construct it from. 
-        If None, pd.DatetimeIndex is constructed from the tstart/tstop/timestep_min metadata attrs of the comp object. 
+        If None, pd.DatetimeIndex is constructed from the tstart/tstop/timestep metadata attrs of the comp object. 
         Only allowed/relevant for component dataframes with single-level columns (single period). The default is None.
+    timestep : str
+        Only allowed/relevant for component dataframes with multi-level columns (different periods). 
+        The string is parsed with pandas.tseries.frequencies.to_offset(). The default is None.
 
     Returns
     -------
@@ -507,13 +508,12 @@ def prediction(comp, timestep_min=None, times=None):
     
     if hasattr(comp.columns,"levels"):
         logger.info('prediction() per period due to levels in component dataframe columns')
-        if timestep_min is None:
-            raise TypeError("prediction() per period, so 'timestep_min' argument should not be None")
+        if timestep is None:
+            raise TypeError("prediction() per period, so 'timestep' argument should not be None")
         if times is not None:
             raise TypeError("prediction() per period, so 'times' argument not allowed")
-        # convert timestep_min to tstep of proper type (tstart/tstop are dummies here)
-        times_slice = slice("1900-01-01", "1900-01-01", timestep_min)
-        _, _, tstep = get_tstart_tstop_tstep(times_slice)
+        # convert timestep to tstep of proper type
+        tstep = pd.tseries.frequencies.to_offset(timestep)
         
         ts_periods_dt = comp.columns.levels[1]
         ts_periods_strlist = [str(x) for x in ts_periods_dt]
@@ -523,12 +523,12 @@ def prediction(comp, timestep_min=None, times=None):
             logger.info(f'generating prediction {period_dt} of sequence {ts_periods_strlist}')
             comp_oneyear = comp.loc[:,(slice(None),period_dt)]
             comp_oneyear.columns = comp_oneyear.columns.droplevel(1)
-            if period_dt.freqstr in ['A-DEC']: #year frequency
-                tstart = dt.datetime(period_dt.year,1,1)
-                tstop = dt.datetime(period_dt.year+1,1,1)-dt.timedelta(minutes=timestep_min)
+            if period_dt.freqstr in ['A-DEC','Y-DEC']: #year frequency
+                tstart = pd.Timestamp(period_dt.year,1,1)
+                tstop = pd.Timestamp(period_dt.year+1,1,1) - tstep.delta
             elif period_dt.freqstr in ['M']: #month frequency
-                tstart = period_dt.to_timestamp().to_pydatetime()
-                tstop = period_dt.to_timestamp().to_pydatetime()+dt.timedelta(days=period_dt.days_in_month)-dt.timedelta(minutes=timestep_min)
+                tstart = period_dt.to_timestamp()
+                tstop = period_dt.to_timestamp() + pd.Timedelta(days=period_dt.days_in_month) - tstep.delta
             else:
                 raise Exception(f'unknown freqstr: {period_dt.freqstr}')
             # generate date range and do prediction
@@ -538,12 +538,14 @@ def prediction(comp, timestep_min=None, times=None):
         ts_prediction = pd.concat(ts_prediction_perperiod_list)
     else:
         logger.info('prediction() atonce')
-        if timestep_min is not None:
-            raise TypeError("prediction() atonce, so 'timestep_min' argument not allowed")
+        if timestep is not None:
+            raise TypeError("prediction() atonce, so 'timestep' argument not allowed")
         if times is None:
             raise TypeError("prediction() atonce, so 'times' argument should not be None")
         if isinstance(times,slice):
-            tstart, tstop, tstep = get_tstart_tstop_tstep(times)
+            tstart = pd.Timestamp(times.start)
+            tstop = pd.Timestamp(times.stop)
+            tstep = pd.tseries.frequencies.to_offset(times.step)
             times = pd.date_range(start=tstart, end=tstop, freq=tstep, unit="us")
         ts_prediction = prediction_singleperiod(comp=comp, times=times, hatyan_settings=hatyan_settings)