+++ tests/examples/longtimeseries_analysis.py
@@ -70,7 +70,7 @@ for current_station in selected_stations:
         #fig.savefig('components_%s_4Y.png'%(current_station))
         
         #prediction and validation
-        times_pred = slice(dt.datetime(2008,1,1),dt.datetime(2010,1,1), 10)
+        times_pred = slice(dt.datetime(2008,1,1),dt.datetime(2010,1,1), "10min")
         ts_prediction = hatyan.prediction(comp=comp_frommeasurements, times=times_pred)
         fig, (ax1,ax2) = hatyan.plot_timeseries(ts=ts_prediction, ts_validation=ts_measurements)
         ax1.set_ylim(-2.5,2.5)