+++ tests/examples/numbering_FEWS_PG.py
@@ -44,8 +44,8 @@ if analyse_ts_bool:
     
 else:
     COMP_merged = hatyan.read_components(filename=file_comp)
-    times_pred_2019 = slice(dt.datetime(2019,1,1),dt.datetime(2019,12,31,23,50), 10)
-    times_pred_2020 = slice(dt.datetime(2020,1,1),dt.datetime(2020,12,31,23,50), 10)
+    times_pred_2019 = slice(dt.datetime(2019,1,1),dt.datetime(2019,12,31,23,50), "10min")
+    times_pred_2020 = slice(dt.datetime(2020,1,1),dt.datetime(2020,12,31,23,50), "10min")
     ts_prediction_fromcomp_2019 = hatyan.prediction(comp=COMP_merged, times=times_pred_2019)
     ts_prediction_fromcomp_2020 = hatyan.prediction(comp=COMP_merged, times=times_pred_2020)
     ts_prediction_fromcomp = pd.concat([ts_prediction_fromcomp_2019,ts_prediction_fromcomp_2020],axis=0)