+++ tests/examples/predictie_2019_19Ycomp4Ydia.py
@@ -75,7 +75,7 @@ for current_station in selected_stations:
     
     file_data_compvali = os.path.join(dir_testdata,'predictie2019',f'{current_station}_ana.txt')
     
-    times_pred = slice(dt.datetime(2019,1,1),dt.datetime(2020,1,1), 10)
+    times_pred = slice(dt.datetime(2019,1,1),dt.datetime(2020,1,1), "10min")
 
     file_data_predvali = os.path.join(dir_testdata,'predictie2019',f'{current_station}_pre.txt')
     file_data_predvaliHWLW = os.path.join(dir_testdata,'predictie2019',f'{current_station}_ext.txt')
@@ -120,15 +120,15 @@ for current_station in selected_stations:
         ts_ext_validation = hatyan.read_dia(filename=file_data_predvaliHWLW, station=current_station)
     else:
         ts_ext_validation = None
-    hatyan.write_dia(ts=ts_prediction, filename='prediction_%im_%s.dia'%(times_pred.step,current_station))
+    hatyan.write_dia(ts=ts_prediction, filename='prediction_%s_%s.dia'%(times_pred.step,current_station))
     ts_ext_prediction = hatyan.calc_HWLW(ts=ts_prediction)
-    hatyan.write_dia(ts=ts_ext_prediction, filename='prediction_HWLW_%im_%s.dia'%(times_pred.step, current_station))
+    hatyan.write_dia(ts=ts_ext_prediction, filename='prediction_HWLW_%s_%s.dia'%(times_pred.step, current_station))
     fig, (ax1,ax2) = hatyan.plot_timeseries(ts=ts_prediction, ts_ext=ts_ext_prediction, ts_ext_validation=ts_ext_validation)
-    fig.savefig('prediction_%im_%s_HWLW'%(times_pred.step, current_station))
+    fig.savefig('prediction_%s_%s_HWLW'%(times_pred.step, current_station))
     fig, (ax1,ax2) = hatyan.plot_timeseries(ts=ts_prediction, ts_validation=ts_validation)
-    fig.savefig('prediction_%im_%s_validation'%(times_pred.step, current_station))
+    fig.savefig('prediction_%s_%s_validation'%(times_pred.step, current_station))
     fig, (ax1,ax2) = hatyan.plot_timeseries(ts=ts_prediction, ts_validation=ts_measurements_group0)
-    fig.savefig('prediction_%im_%s_measurements'%(times_pred.step, current_station))
+    fig.savefig('prediction_%s_%s_measurements'%(times_pred.step, current_station))
 
 print('\nthese %i stations were requested for processing:\n%s'%(len(selected_stations),selected_stations))
 print('\nthese %i stations were not processed because there is no ana/comp dataset available:\n%s'%(len(stats_noana),stats_noana))