+++ tests/examples/predictie_2019_frommergedcomp.py
@@ -47,7 +47,7 @@ for current_station in selected_stations:
 
     file_data_compvali = os.path.join(dir_testdata,'predictie2019','%s_ana.txt'%(current_station))
     
-    times_pred = slice(dt.datetime(2019,1,1),dt.datetime(2020,1,1), 10)
+    times_pred = slice(dt.datetime(2019,1,1),dt.datetime(2020,1,1), "10min")
     
     file_data_predvali = os.path.join(dir_testdata,'predictie2019','%s_pre.txt'%(current_station))
     file_data_predvaliHWLW = os.path.join(dir_testdata,'predictie2019','%s_ext.txt'%(current_station))
@@ -64,13 +64,13 @@ for current_station in selected_stations:
     ts_prediction = hatyan.prediction(comp=COMP_merged, times=times_pred)
     ts_validation = hatyan.read_dia(filename=file_data_predvali, station=current_station)
     #ts_ext_validation = hatyan.read_dia(filename=file_data_predvaliHWLW, station=current_station)
-    hatyan.write_dia(ts=ts_prediction, filename='prediction_%im_%s.dia'%(times_pred.step,current_station))
+    hatyan.write_dia(ts=ts_prediction, filename='prediction_%s_%s.dia'%(times_pred.step,current_station))
     #ts_ext_prediction = hatyan.calc_HWLW(ts=ts_prediction)
-    #hatyan.write_dia(ts=ts_ext_prediction, filename='prediction_HWLW_%im_%s.dia'%(times_pred.step, current_station))
+    #hatyan.write_dia(ts=ts_ext_prediction, filename='prediction_HWLW_%s_%s.dia'%(times_pred.step, current_station))
     #fig, (ax1,ax2) = hatyan.plot_timeseries(ts=ts_prediction, ts_ext=ts_ext_prediction, ts_ext_validation=ts_ext_validation)
-    #fig.savefig('prediction_%im_%s_HWLW'%(times_pred.step, current_station))
+    #fig.savefig('prediction_%s_%s_HWLW'%(times_pred.step, current_station))
     fig, (ax1,ax2) = hatyan.plot_timeseries(ts=ts_prediction, ts_validation=ts_validation)
-    fig.savefig('prediction_%im_%s_validation'%(times_pred.step, current_station))
+    fig.savefig('prediction_%s_%s_validation'%(times_pred.step, current_station))
 
 print('\nthese %i stations were requested for processing:\n%s'%(len(selected_stations),selected_stations))
 print('\nthese %i stations were not processed because there is no ana/comp dataset available:\n%s'%(len(stats_noana),stats_noana))