+++ docs/whats-new.md
@@ -23,6 +23,7 @@
 - made all read/write timeserie function names consistent by dropping "ts" in [#265](https://github.com/Deltares/hatyan/pull/265)
 - use prediction settings from `comp` argument instead of settings arguments in `hatyan.prediction()` [#263](https://github.com/Deltares/hatyan/issues/263)
 - re-introduce clear arguments for `hatyan.analysis()` in [#270](https://github.com/Deltares/hatyan/pull/270)
+- dropped support for `ìnt` timesteps in slices and in `hatyan.prediction()` [#288](https://github.com/Deltares/hatyan/pull/288)
 
 ### Deprecated
 - hatyan ddl functions `get_DDL_catalog()`, `get_DDL_queryserver()`, `get_DDL_data()`, `get_DDL_stationmetasubset()` in [#206](https://github.com/Deltares/hatyan/pull/206)